+++ intern/cycles/kernel/svm/svm.h
@@ -98,7 +98,7 @@ CCL_NAMESPACE_BEGIN
 #endif
 
 /* Main Interpreter Loop */
-template<uint node_feature_mask, ShaderType type, typename ConstIntegratorGenericState>
+template<uint64_t node_feature_mask, ShaderType type, typename ConstIntegratorGenericState>
 ccl_device void svm_eval_nodes(KernelGlobals kg,
                                ConstIntegratorGenericState state,
                                ccl_private ShaderData *sd,
