+++ intern/cycles/kernel/types.h
@@ -1370,7 +1370,7 @@ struct KernelSceneTime {
 
 struct ccl_align(16) KernelData {
   /* Features and limits. */
-  uint kernel_features;
+  uint64_t kernel_features;
   uint max_closures;
   uint max_shaders;
   uint volume_stack_size;
