+++ intern/cycles/scene/scene.h
@@ -267,7 +267,7 @@ class Scene : public NodeOwner {
   void free_memory(bool final);
 
   bool kernels_loaded;
-  uint loaded_kernel_features;
+  uint64_t loaded_kernel_features;
 
   void update_kernel_features();
