+++ intern/cycles/scene/shader.cpp
@@ -836,9 +836,9 @@ void ShaderManager::add_default(Scene *scene)
   }
 }
 
-uint ShaderManager::get_graph_kernel_features(ShaderGraph *graph)
+uint64_t ShaderManager::get_graph_kernel_features(ShaderGraph *graph)
 {
-  uint kernel_features = 0;
+  uint64_t kernel_features = 0;
 
   for (ShaderNode *node : graph->nodes) {
     kernel_features |= node->get_feature();
@@ -859,9 +859,9 @@ uint ShaderManager::get_graph_kernel_features(ShaderGraph *graph)
   return kernel_features;
 }
 
-uint ShaderManager::get_kernel_features(Scene *scene)
+uint64_t ShaderManager::get_kernel_features(Scene *scene)
 {
-  uint kernel_features = KERNEL_FEATURE_NODE_BSDF | KERNEL_FEATURE_NODE_EMISSION;
+  uint64_t kernel_features = KERNEL_FEATURE_NODE_BSDF | KERNEL_FEATURE_NODE_EMISSION;
   for (int i = 0; i < scene->shaders.size(); i++) {
     Shader *shader = scene->shaders[i];
     if (!shader->reference_count()) {
