+++ intern/cycles/scene/shader.h
@@ -214,7 +214,7 @@ class ShaderManager {
   static void add_default(Scene *scene);
 
   /* Selective nodes compilation. */
-  uint get_kernel_features(Scene *scene);
+  uint64_t get_kernel_features(Scene *scene);
 
   float linear_rgb_to_gray(const float3 c);
   float3 rec709_to_scene_linear(const float3 c);
@@ -270,7 +270,7 @@ class ShaderManager {
 
   void compute_thin_film_table(const Transform &xyz_to_rgb);
 
-  uint get_graph_kernel_features(ShaderGraph *graph);
+  uint64_t get_graph_kernel_features(ShaderGraph *graph);
 
   virtual void device_update_specific(Device *device,
                                       DeviceScene *dscene,
